+++ plasmapy/formulary/parameters.py
@@ -23,18 +23,18 @@ import numpy as np
 from astropy.constants.si import e, eps0, k_B, mu0
 from typing import Optional, Union
 
-from plasmapy import particles
 from plasmapy.particles import Particle
 from plasmapy.utils.decorators import deprecated, validate_quantities
 from plasmapy.utils.exceptions import PlasmaPyFutureWarning
 
-from plasmapy.formulary import dimensionless, frequencies, lengths, speeds  # noqa
+from plasmapy.formulary import dimensionless, frequencies, lengths, misc, speeds  # noqa
 
 
 __all__ += (
     dimensionless.__all__.copy()
     + frequencies.__all__.copy()
     + lengths.__all__.copy()
+    + misc.__all__.copy()
     + speeds.__all__.copy()
     + __aliases__
     + __lite_funcs__
@@ -44,6 +44,7 @@ __aliases__ += (
     dimensionless.__aliases__.copy()
     + frequencies.__aliases__.copy()
     + lengths.__aliases__.copy()
+    + misc.__aliases__.copy()
     + speeds.__aliases__.copy()
 )
 
@@ -98,6 +99,7 @@ funcs_to_deprecate_wrap = [  # (module_name, func_name)
     ("lengths", "rhoc_"),
     ("lengths", "inertial_length"),
     ("lengths", "cwp_"),
+    ("misc", "_grab_charge"),
     ("speeds", "Alfven_speed"),
     ("speeds", "va_"),
     ("speeds", "ion_sound_speed"),
@@ -122,36 +124,6 @@ for modname, name in funcs_to_deprecate_wrap:
     )(getattr(globals()[f"{modname}"], name))
 
 del modname, name
-
-
-def _grab_charge(ion: Particle, z_mean=None):
-    """
-    Merge two possible inputs for particle charge.
-
-    Parameters
-    ----------
-    ion : `~plasmapy.particles.particle_class.Particle`
-        a string representing a charged particle, or a Particle object.
-
-    z_mean : `float`
-        An optional float describing the average ionization of a particle
-        species.
-
-    Returns
-    -------
-    float
-        if ``z_mean`` was passed, ``z_mean``, otherwise, the charge number
-        of ``ion``.
-
-    """
-    if z_mean is None:
-        # warnings.warn("No z_mean given, defaulting to atomic charge",
-        #               PhysicsWarning)
-        Z = particles.charge_number(ion)
-    else:
-        # using average ionization provided by user
-        Z = z_mean
-    return Z
 
 
 @validate_quantities(
